DynamoDB get put and delete hard coded working.
