+++ src/TM470Serverless-Lambda/TM470GetStations.js
@@ -5,16 +5,18 @@ exports.handler = async(event, context) => {
 
     const documentClient = new AWS.DynamoDB.DocumentClient();
 
-
     let responseBody  = "";
     let statusCode = 0;
+
 
     const params = {
         TableName: "Stations"
     };
+
+    
     try {
         const data = await documentClient.scan(params).promise();
-        responseBody = JSON.stringify(data.Items);
+        responseBody = JSON.stringify(data);
         statusCode = 200;
     } catch (err) {
         responseBody = 'unable to get stations: ${err}';
